LMSProject/requirements.txt
Django==2.1.2
psycopg2
requests
oauth2
social-auth-app-django
pyjwt
python-jose
six
python-dotenv